Added rudimentary distribution generation to vegetation (mostly untested, but it creates a sensible image)
Distributions are currently generated completely randomly based on the number of species in it.
A field "distribution_type" has been added to SpeciesRepresentation, so that in the future, certain types of plants will be distributed differently in the graphic.

A complete distribution for the whole phytocoenosis is created, as well as separate distribution graphics for each layer.