Orthos are retrieved using tile.get_tile()
For this, the file ending can now (optionally) be specified in tiles.get_tile(). (orthos are currently the only tiles which use .jpg instead of .png)

A downside of this change is that ortho tiles are now never fetched from the internet if they don't exist. We may want to re-add this at some point, before the get_tile call.