Workaround for failing tile requests due to folder being created twice
This workaround makes two concurrent raster requests which generate a new tile in the same new folder not fail. Since the client is currently guaranteed to never send the exact same raster request twice, this should fix the current problem. However, concurrent identical requests should be handled correctly in the future as well.