Changes in vegetations model to reflect new definitions
Some fields were removed since they're not needed: distribution_graphic_path is chosen via filesystem path, and the parameters max_height, max_slope, min_height, min_slope are not needed since Phytocoenosis are placed based on Invekos data anyways.

The distribution_density definition was changed to be occurances/m² since that's the preferred way to define it. The algorithm still needs to be changed to accurately reflect that.