Untested draft of line segment request
The request works similarly to dynamic assets: All lines within a given
radius are returned. I think that this is the most universal approach,
and we may be able to unify all assets (static, dynamic, linear) with
this type of retrieval later, if it works well. Since it doesn't use
Tiles, we also get rid of the extra step of assigning lines to tiles.