Remove incomplete entries in vegetation csv
Due to peculiarities in the spreadsheet or the CSV conversion, it's
possible that we get speciesRepresentations with no billboard in the
command. These are deleted and not included in the generated fixture. A
warning is issued since it's possible that it was a data entry error and
a phytocoenosis is actually dependent on it.