Add request to get AssetPositions near given point
The AssetPositions can be filtered either by Asset ID or by AssetType
ID.

The display_radius is used as the maximum distance of the AssetPosition
to the given point. If 0, all are returned.

This request should improve performance (if there are many Assets, only
the relevant ones are returned, not all). It will also allow us to unify
how Buildings and dynamic assets like Windmills are handled.